refactor: simplify governance contract interactions
Moves proposal ID extraction logic to governance contract
Removes separate events file
Updates contract interaction methods to use new governance proposer contract methods
Simplifies address retrieval and contract initialization
