+++ yarn-project/ethereum/src/contracts/governance.ts
@@ -7,15 +7,15 @@ import {
   type EncodeFunctionDataParameters,
   type GetContractReturnType,
   type Hex,
+  type Log,
   encodeFunctionData,
   getContract,
+  parseEventLogs,
 } from 'viem';
 
 import type { L1ContractAddresses } from '../l1_contract_addresses.js';
 import { L1TxUtils } from '../l1_tx_utils.js';
 import type { ViemPublicClient, ViemWalletClient } from '../types.js';
-import { extractProposalIdFromLogs } from './events.js';
-import { GovernanceProposerContract } from './governance_proposer.js';
 
 export type L1GovernanceContractAddresses = Pick<
   L1ContractAddresses,
@@ -32,6 +32,19 @@ export enum ProposalState {
   Executed,
   Dropped,
   Expired,
+}
+
+export function extractProposalIdFromLogs(logs: Log[]): bigint {
+  const parsedLogs = parseEventLogs({
+    abi: GovernanceAbi,
+    logs: logs,
+    eventName: 'Proposed',
+  });
+
+  if (parsedLogs.length === 0) {
+    throw new Error('Proposal log not found');
+  }
+  return parsedLogs[0].args.proposalId;
 }
 
 export class GovernanceContract {
@@ -53,23 +66,8 @@ export class GovernanceContract {
     return EthAddress.fromString(this.publicGovernance.address);
   }
 
-  public async getProposer() {
-    const governanceProposerAddress = EthAddress.fromString(await this.publicGovernance.read.governanceProposer());
-    return new GovernanceProposerContract(this.publicClient, governanceProposerAddress.toString());
-  }
-
-  public async getGovernanceAddresses(): Promise<L1GovernanceContractAddresses> {
-    const governanceProposer = await this.getProposer();
-    const [rollupAddress, registryAddress] = await Promise.all([
-      governanceProposer.getRollupAddress(),
-      governanceProposer.getRegistryAddress(),
-    ]);
-    return {
-      governanceAddress: this.address,
-      rollupAddress,
-      registryAddress,
-      governanceProposerAddress: governanceProposer.address,
-    };
+  public async getGovernanceProposerAddress() {
+    return EthAddress.fromString(await this.publicGovernance.read.governanceProposer());
   }
 
   public getConfiguration() {
