+++ yarn-project/ethereum/src/contracts/governance_proposer.ts
@@ -7,7 +7,7 @@ import { type GetContractReturnType, type Hex, type TransactionReceipt, encodeFu
 import type { GasPrice, L1TxRequest, L1TxUtils } from '../l1_tx_utils.js';
 import type { ViemPublicClient } from '../types.js';
 import { type IEmpireBase, encodeVote } from './empire_base.js';
-import { extractProposalIdFromLogs } from './events.js';
+import { extractProposalIdFromLogs } from './governance.js';
 
 export class GovernanceProposerContract implements IEmpireBase {
   private readonly proposer: GetContractReturnType<typeof GovernanceProposerAbi, ViemPublicClient>;
